Refactor cairo threads
